pyproject.toml
[tool.poetry]
name = "pyzeebe"
version = "4.0.0rc7"
description = "Zeebe client api"
authors = ["Jonatan Martens <jonatanmartenstav@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/camunda-community-hub/pyzeebe"
repository = "https://github.com/camunda-community-hub/pyzeebe"
documentation = "https://pyzeebe.readthedocs.io/en/latest"
keywords = ["zeebe", "workflow", "workflow-engine"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
packages = [
    { include = "pyzeebe" },
]

[tool.poetry.dependencies]
python = "^3.9"
oauthlib = "^3.1.0"
requests-oauthlib = ">=1.3.0,<3.0.0"
zeebe-grpc = "^8.4.0"
typing-extensions = "^4.11.0"
anyio = "^4.6.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4,<9.0"
pytest-asyncio = "^0.21.1"
pytest-grpc = "^0.8.0"
pytest-mock = "^3.11.1"
pylint = ">=2.17.5,<4.0.0"
black = "^24.0.0"
mypy = "^1.10.0"
coveralls = "^3.3.1"
responses = ">=0.23.2,<0.26.0"
sphinx-rtd-theme = ">=3.0.0,<3.1.0"
sphinx = ">=6,<8"

[tool.poetry.group.stubs.dependencies]
types-oauthlib = "^3.1.0"
types-requests-oauthlib = ">=1.3.0,<3.0.0"

[tool.mypy]
python_version = "3.9"
packages = ["pyzeebe"]
strict = true

[[tool.mypy.overrides]]
module = [
  "grpc", 
  "zeebe_grpc.gateway_pb2", 
  "zeebe_grpc.gateway_pb2_grpc", 
]
ignore_missing_imports = true

[tool.pylint.master]
max-line-length = 120
disable = ["C0114", "C0115", "C0116"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "TID", # flake8-tidy-imports
    "T20", # flake8-print
    "ASYNC", # flake8-async
    "FA",  # flake8-future-annotations
]
ignore = [
    "E501",   # line too long, handled by black
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
